Add integration test

+++ src/main/java/com/fibo/model/FibonacciResult.java
@@ -1,5 +1,7 @@
 package com.fibo.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,6 +11,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class FibonacciResult {
 
-	private Integer value;
+	private BigDecimal value;
 	
 }
